Licensing: GPL -> LGPL  and clarifications
The GPL license without further exceptions, would not have been a good license
choice, as no one but me would have been able to distribute the program, as GPL
conditions and the nature of a Minecraft mod clash.
I added the neccessary exceptions and relicensed to LGPL. This should now
(hopefully) work the way I want it to and allow others to fork this mod if they
wish to do so.